In php 7, zend_parse_parameters specs changed. 's' expects size_t now instead of int.

+++ php_jsmin.c
@@ -100,7 +100,7 @@ ZEND_GET_MODULE(jsmin)
 PHP_FUNCTION(jsmin)
 {
 	char *javascript;
-	int javascript_len;
+	size_t javascript_len;
 	jsmin_obj *jmo;
 
 	zval *ret_code = NULL;
